HOLY CRAP We have an entire command structure. And it's got some fun ideas. The system is built around four main components: Triggers, Dispatcher, Commands, and Authorizers. Dispatcher is a singleton that stores commands along with their associated Authorizers in a list. Whenever an event is emitted from a Trigger, the Dispatcher checks with the corresponding Authorizer to see if the command should be executed. There’s a wide variety of Authorizers for different use cases.
Triggers simply notify the Dispatcher that something has happened within a given context. There are several types of triggers, with the two main ones being TimeTrigger and SlashBinding. TimeTriggers schedule events via APScheduler, which is managed by its own singleton elsewhere. Bindings represent human interaction with triggers; currently, SlashBinding is the only example, requiring a name and description for Discord’s command UI.

Commands are classes that accept arguments and a CommandData object, then define an execute method based on that data. CommandBuilders are the only approved way to instantiate Commands. This design exists so that @on_ready can process CommandData, which must be defined at runtime.

Commands now have convenient query tools: you can attach a View to messages and use the Container object to store the user’s selected result. Views can be awaited, and all you need is to pass a consumer function to the view along with a dictionary mapping display strings to result values. This makes command interactions simple and flexible.

Now, EventContext and EventType and CommandData are designed to be changed. All arguments need to default to None. Other classes shouldn't ever touch those new arguments, so it should be perfectly fine.

Things that need changed:
I need to find a good way to make each Authorizer a Boolean Supplier
I need to get rid of the ABC in the file system, but that's for readability
I need to add more Queries, such as the multi-select list option
I need to move discord's main commanding logic outside of @on_ready.